Edit to Sprint Challenge Directions
Formatting and Documentation are very crucial for time-sensitive projects. Foresight and consideration on part of the author should be used when knowing which information to highlight and which information would be extraneous and not helpful to the direct product of the end goal. 
Implementing formatting helps students to achieve the goal of the challenge within the time constraints, while also helping students who have mental, psychological, and physical disabilities like ADHD, GAD, and colorblindness.